refactor(096-W4): models/mod.rs 按域拆分——todo/execution/executor 三文件（2108→936 行）
- Move Function 逐字搬迁：todo.rs（39 段 657 行：TodoStatus/Todo/ComputedBucket/分页族/
  请求 DTO/模板族/备份 DTO/伪 ID 与占位符工具）、execution.rs（29 段：ExecutionStatus/
  ExecutionRecord/ParsedLogEntry/Dashboard 统计族/UTC 时间工具）、executor.rs（8 段：
  ExecutorType/配置/检测结果族）
- 聚合根零破坏：pub use todo::*/execution::*/executor::* 保持 crate::models::Xxx 外部
  引用路径零改动；测试模块全部留 mod.rs（use super::* 经聚合根编译不感知）
- 跨域引用修正：Todo/ExecutionUsage 经 use 引入；default_true/default_trigger_type
  提 pub(crate) + serde default 属性改全路径
- 守恒核验：pub 项 87→89（+2 为工具函数可见性提升）；全量 1677 测试过、clippy 零告警
