+++ backend/src/models/mod.rs
@@ -3,341 +3,24 @@ use serde::{Deserialize, Serialize};
 pub mod loop_;
 pub use loop_::*;
 
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum TodoStatus {
-    Pending,
-    InProgress,
-    Running,
-    Completed,
-    Failed,
-    Cancelled,
-}
+// 096-W4-2：按域拆分四文件（todo/execution/loop/executor）。
+// pub use 聚合保持外部引用路径（crate::models::Xxx）零改动。
+pub mod execution;
+pub mod executor;
+pub mod todo;
+pub use execution::*;
+pub use executor::*;
+pub use todo::*;
 
-impl TodoStatus {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::Pending => "pending",
-            Self::InProgress => "in_progress",
-            Self::Running => "running",
-            Self::Completed => "completed",
-            Self::Failed => "failed",
-            Self::Cancelled => "cancelled",
-        }
-    }
-}
 
-impl std::fmt::Display for TodoStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.as_str())
-    }
-}
 
-impl std::str::FromStr for TodoStatus {
-    type Err = String;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "pending" => Ok(Self::Pending),
-            "in_progress" => Ok(Self::InProgress),
-            "running" => Ok(Self::Running),
-            "completed" => Ok(Self::Completed),
-            "failed" => Ok(Self::Failed),
-            "cancelled" => Ok(Self::Cancelled),
-            _ => Err(format!("unknown status: {}", s)),
-        }
-    }
-}
 
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum ExecutionStatus {
-    Running,
-    Success,
-    Failed,
-}
 
-impl ExecutionStatus {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::Running => "running",
-            Self::Success => "success",
-            Self::Failed => "failed",
-        }
-    }
-}
 
-impl std::fmt::Display for ExecutionStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.as_str())
-    }
-}
 
-impl std::str::FromStr for ExecutionStatus {
-    type Err = String;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "running" => Ok(Self::Running),
-            "success" => Ok(Self::Success),
-            "failed" => Ok(Self::Failed),
-            _ => Err(format!("unknown execution status: {}", s)),
-        }
-    }
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Todo {
-    pub id: i64,
-    pub title: String,
-    pub prompt: String,
-    pub status: TodoStatus,
-    pub created_at: String,
-    pub updated_at: String,
-    #[serde(default)]
-    pub tag_ids: Vec<i64>,
-    #[serde(default)]
-    pub executor: Option<String>,
-    #[serde(default)]
-    pub scheduler_enabled: bool,
-    #[serde(default)]
-    pub scheduler_config: Option<String>,
-    #[serde(default)]
-    pub scheduler_timezone: Option<String>,
-    #[serde(default)]
-    pub scheduler_next_run_at: Option<String>,
-    #[serde(default)]
-    pub task_id: Option<String>,
-    /// 工作空间目录路径（cwd，仅后端内部使用，不通过 API 暴露给前端）。
-    /// 业务层（前端 / CLI / sync）只通过 `workspace_id` 标识工作空间，
-    /// path 字段保留供 executor_service / worktree 等需要 cwd 的子系统使用。
-    #[serde(default)]
-    pub workspace_path: Option<String>,
-    /// 所属工作空间 ID（project_directories.id），唯一键。
-    /// 业务层（前端 / CLI / API）统一以此作为工作空间标识符。
-    #[serde(default)]
-    pub workspace_id: Option<i64>,
-    #[serde(default)]
-    pub webhook_enabled: bool,
-    #[serde(default)]
-    pub acceptance_criteria: Option<String>,
-    /// 0=normal, 1=reviewer_template（已废弃：评审模板已迁出至 review_templates 表）,
-    /// 2=review_instance（评审实例）.
-    #[serde(default)]
-    pub todo_type: i32,
-    /// review_instance 关联到被评审的原 todo; 其它类型为 None.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub parent_todo_id: Option<i64>,
-    /// review_instance 关联到生成它的 review_template; 其它类型为 None.
-    /// NULL/NONE 可能是 V15 之前的迁移产物.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub review_template_id: Option<i64>,
-    /// 是否在执行完成后自动派生一个评审 todo. 只对 normal 类型有意义.
-    #[serde(default = "default_true")]
-    pub auto_review_enabled: bool,
-    /// Action 类型标记（如 "title_optimize"、"prompt_optimize"）。
-    /// 与 action_key 配合，由 /api/actions/execute 用于查找或自动创建 action 模板 todo。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub action_type: Option<String>,
-    /// Action 键值，与 action_type 配合唯一标识一个 action 模板 todo。
-    /// 由 /api/actions/execute 用于查找或自动创建 action 模板 todo。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub action_key: Option<String>,
-    /// 归档时间戳（UTC 字符串）。None=未归档，参与事项中心日常分类；
-    /// Some=已归档，进入「已归档」分类，从日常视图隐藏但数据保留。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub archived_at: Option<String>,
-    /// 专家/团队名称（WorkBuddy plugin.json 中的 name 字段）。
-    /// 执行时自动加载对应的 Agent MD 和 Skills 注入 prompt。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub expert_name: Option<String>,
-    /// 任务级指定的执行模型（覆盖 executor.default_model）。
-    /// None = 未指定，执行时回退到执行器默认模型；执行器也未指定则不传 --model。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub model: Option<String>,
-    /// 事项级技能名列表（需求 055）。工艺安装时从环节 skills 写入；
-    /// 执行时以 `/skill-name` 逐行注入 prompt 尾部，由执行器 CLI 解析。
-    /// `#[serde(default)]` 保证旧客户端/旧数据反序列化不受影响。
-    #[serde(default)]
-    pub skills: Vec<String>,
-}
 
-/// 事项中心的五类驱动分类（computed_bucket）。
-///
-/// 这是运行时由底层事实字段推导的返回值，不落库。
-/// 推导规则见 `compute_bucket`，优先级：已归档 > Loop 驱动 > 时间驱动 > 事件驱动 > 手动触发。
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ComputedBucket {
-    /// 手动触发：兜底分类，未归档且无调度/无 Webhook/未被 Loop 引用。
-    Manual,
-    /// 时间驱动：scheduler_config 非空（scheduler_enabled 仅表启停）。
-    TimeDriven,
-    /// 事件驱动：webhook_enabled=true 且无调度配置且未被 Loop 引用。
-    EventDriven,
-    /// Loop 驱动：被启用的 loop_steps 引用（used_by_loop_step_count > 0）。
-    LoopDriven,
-    /// 已归档：archived_at 非空，优先级最高。
-    Archived,
-}
 
-impl ComputedBucket {
-    /// 从查询参数解析分类。None/空字符串=不过滤（返回全部）。
-    /// 不区分大小写，与 serde 的 snake_case 序列化对齐。
-    pub fn parse_query(s: &str) -> Option<Self> {
-        match s.trim().to_ascii_lowercase().as_str() {
-            "manual" => Some(Self::Manual),
-            "time_driven" => Some(Self::TimeDriven),
-            "event_driven" => Some(Self::EventDriven),
-            "loop_driven" => Some(Self::LoopDriven),
-            "archived" => Some(Self::Archived),
-            _ => None,
-        }
-    }
-}
-
-/// 由底层事实字段推导事项的主分类（computed_bucket）。
-///
-/// 优先级（设计文档）：已归档 > Loop 驱动 > 时间驱动 > 事件驱动 > 手动触发。
-/// 纯函数、无 IO，便于单测覆盖各组合分支。
-///
-/// - `archived_at` 非空 → 已归档（最高优先级，用户明确希望隐藏）
-/// - 否则被 Loop 引用（count>0）→ Loop 驱动（已成流程结构一部分）
-/// - 否则 scheduler_config 非空 → 时间驱动（注意：scheduler_enabled 仅表启停，不决定是否时间驱动）
-/// - 否则 webhook_enabled → 事件驱动
-/// - 否则 → 手动触发（兜底）
-pub fn compute_bucket(
-    archived_at: Option<&str>,
-    used_by_loop_step_count: i64,
-    scheduler_config: Option<&str>,
-    webhook_enabled: bool,
-) -> ComputedBucket {
-    if archived_at.is_some() {
-        return ComputedBucket::Archived;
-    }
-    if used_by_loop_step_count > 0 {
-        return ComputedBucket::LoopDriven;
-    }
-    if scheduler_config.is_some() {
-        return ComputedBucket::TimeDriven;
-    }
-    if webhook_enabled {
-        return ComputedBucket::EventDriven;
-    }
-    ComputedBucket::Manual
-}
-
-/// 事项中心列表项：在 Todo 之上附加运行时推导/聚合字段。
-///
-/// 附加字段由 handler 层批量补算（loop 引用计数、最近一次执行记录），
-/// 普通 `get_todos` 路径不返回这些字段，因此独立成 DTO 而非塞进 Todo。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TodoCenterItem {
-    /// 内联 Todo 全部字段，保持响应扁平（设计文档示例即扁平结构）。
-    #[serde(flatten)]
-    pub todo: Todo,
-    /// 运行时推导的主分类，不落库。
-    pub computed_bucket: ComputedBucket,
-    /// 被启用 loop_steps 引用的次数（COUNT ... WHERE enabled=1 GROUP BY todo_id）。
-    /// 0=未被任何启用的 Loop 引用。
-    #[serde(default)]
-    pub used_by_loop_step_count: i64,
-    /// 最近一次执行记录的状态（success/failed/running/...），无记录则 None。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub last_execution_status: Option<String>,
-    /// 最近一次执行记录的时间（优先 finished_at，回退 started_at），无记录则 None。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub last_execution_at: Option<String>,
-    /// 引用该事项的启用 Loop 摘要（loop_id + name）。仅 Loop 驱动分类非空，
-    /// 供卡片展示「所属 Loop」并跳转 Loop 详情。空 vec=未被引用或非 Loop 驱动。
-    #[serde(default)]
-    pub referencing_loops: Vec<LoopRefSummary>,
-    /// 连续失败次数：从最近一次执行往前数连续 failed 的条数，遇非 failed 即停。
-    /// 0=最近一次非失败（或无记录）。时间/事件驱动卡片健康展示用。
-    #[serde(default)]
-    pub consecutive_failure_count: i64,
-    /// 最近一次 webhook 触发的时间（trigger_type='webhook' 的最新记录）。
-    /// 仅事件驱动卡片展示「最近触发时间」用；手动「执行一次」不顶替该时间。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub last_webhook_trigger_at: Option<String>,
-    /// 绑定的工作空间斜杠命令（command_type='todo' 绑定该 todo 的第一条）。
-    /// 卡片展示「绑定命令: /xxx」；手动触发 Tab 可据此筛「仅看可命令触发」。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub bound_slash_command: Option<String>,
-}
-
-/// 事项中心服务端分页响应（056）。
-///
-/// `bucket_counts` 供前端分类 Tab 角标：统计口径=应用 search 后、应用 bucket 过滤前，
-/// 与「分页前完成分桶」的旧内存实现语义一致。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TodoCenterPage {
-    pub items: Vec<TodoCenterItem>,
-    pub total: i64,
-    pub page: i64,
-    pub page_size: i64,
-    pub bucket_counts: std::collections::HashMap<String, i64>,
-    /// 当前工作空间内出现过的 action_type 去重列表（卡片墙「来源筛选」下拉数据源）
-    #[serde(default)]
-    pub action_types: Vec<String>,
-}
-
-/// 事项轻量摘要（056）：不含 prompt/acceptance_criteria 大文本字段。
-///
-/// 供看板全量渲染、下拉选择、记录详情补标题等「只需要展示字段」的场景，
-/// 替代过去「整行 Todo 全量拉取」的做法。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TodoBrief {
-    pub id: i64,
-    pub title: String,
-    pub status: TodoStatus,
-    #[serde(default)]
-    pub executor: Option<String>,
-    pub updated_at: String,
-    #[serde(default)]
-    pub archived_at: Option<String>,
-    /// 所属工作空间（拖拽/归属展示用，小字段）
-    #[serde(default)]
-    pub workspace_id: Option<i64>,
-    /// 标签 id 列表（看板标签徽章用；批量查询补算，不逐行 N+1）
-    #[serde(default)]
-    pub tag_ids: Vec<i64>,
-    /// prompt 是否非空（看板「展开 prompt」区块的显示开关，不必传输 prompt 本体）
-    #[serde(default)]
-    pub has_prompt: bool,
-}
-
-/// 事项列表分页响应（056，旧全量接口改造后的统一结构）。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TodoListPage {
-    pub items: Vec<Todo>,
-    pub total: i64,
-    pub page: i64,
-    pub page_size: i64,
-}
-
-/// Loop 引用摘要：事项中心展示「所属环路」与「工艺」两列共用。
-/// - loop_id/loop_name：引用该事项的环路实例。
-/// - process_template_*：该环路所基于的工艺模板（loops.process_template_id → process_templates），
-///   供「工艺」列按 #模板ID-模板名-版本 展示；环路无模板时为 None。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LoopRefSummary {
-    pub loop_id: i64,
-    pub loop_name: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub process_template_id: Option<i64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub process_template_name: Option<String>,
-    /// 工艺版本：优先取环路实例化快照，缺失时由 SQL 回退模板当前版本。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub process_template_version: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Tag {
-    pub id: i64,
-    pub name: String,
-    pub color: String,
-    pub created_at: String,
-}
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct AgentBot {
@@ -381,612 +64,54 @@ impl Default for BotConfig {
     }
 }
 
-fn default_true() -> bool {
+pub(crate) fn default_true() -> bool {
     true
 }
 
-fn default_trigger_type() -> String { "manual".to_string() }
+pub(crate) fn default_trigger_type() -> String { "manual".to_string() }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecutionRecord {
-    pub id: i64,
-    pub todo_id: i64,
-    pub status: ExecutionStatus,
-    pub command: String,
-    pub stdout: String,
-    pub stderr: String,
-    pub result: Option<String>,
-    pub started_at: String,
-    pub finished_at: Option<String>,
-    pub usage: Option<ExecutionUsage>,
-    pub executor: Option<String>,
-    pub model: Option<String>,
-    #[serde(default = "default_trigger_type")]
-    pub trigger_type: String,
-    #[serde(default)]
-    pub pid: Option<i32>,
-    #[serde(default)]
-    pub task_id: Option<String>,
-    #[serde(default)]
-    pub session_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub todo_progress: Option<String>,
-    /// 多 Agent 协作的子 agent 元数据（JSON 字符串 `Vec<AgentRun>`）。
-    /// 透传 entity 的 agent_runs 列，前端自行 parse。与 todo_progress 一样不在此处反序列化。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub agent_runs: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub execution_stats: Option<ExecutionStats>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub resume_message: Option<String>,
-    /// Hook trigger provenance: the source todo that fired this execution.
-    /// `Some` only when `trigger_type` starts with `hook:`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source_todo_id: Option<i64>,
-    /// Snapshot of the source todo's title at trigger time.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source_todo_title: Option<String>,
-    /// User-provided score for this execution's result (0-100, optional).
-    /// Only meaningful on terminal records (success/failed).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub rating: Option<i32>,
-    /// 自动评审时, 评审记录精确指向被评审的"原执行记录"。
-    /// 这条记录的 rating 应被视为对 source_execution_record_id 的评分.
-    /// NULL = 这条记录不是被自动评审的产物.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source_execution_record_id: Option<i64>,
-    /// 这条原执行记录最近一次自动评审的状态.
-    /// 仅在原执行记录上有意义; 评审实例自己的 execution_record 该字段为 NULL.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub last_review_status: Option<String>,
-    /// 这条原执行记录最近一次自动评审 spawn 的 UTC 时间.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub last_reviewed_at: Option<String>,
-    /// issue #643: 本次执行使用的 git worktree 目录。None = 未启用 worktree 或未创建成功。
-    /// 字段语义：仅供"事后排查"，不影响子进程 cwd；auto_cleanup 决定它在执行后是否被删。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub worktree_path: Option<String>,
-    /// 当本次执行是 loop 环节的一部分时，指向 loop_step_executions 表的 id。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub loop_step_execution_id: Option<i64>,
-    /// 已废弃，曾用于环节独立执行。现始终为 None。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub step_id: Option<i64>,
-    /// record 直接归属的 workspace（v89 新增）。归属校验用它，不再经 todo 间接关联。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub workspace_id: Option<i64>,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecutionUsage {
-    pub input_tokens: u64,
-    pub output_tokens: u64,
-    pub cache_read_input_tokens: Option<u64>,
-    pub cache_creation_input_tokens: Option<u64>,
-    pub total_cost_usd: Option<f64>,
-    pub duration_ms: Option<u64>,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecutionStats {
-    pub tool_calls: u64,
-    pub conversation_turns: u64,
-    pub thinking_count: u64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecutionSummary {
-    pub todo_id: i64,
-    pub total_executions: i64,
-    pub success_count: i64,
-    pub failed_count: i64,
-    pub running_count: i64,
-    pub total_input_tokens: u64,
-    pub total_output_tokens: u64,
-    pub total_cache_read_tokens: u64,
-    pub total_cache_creation_tokens: u64,
-    pub total_cost_usd: Option<f64>,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TodoItem {
-    pub id: Option<String>,
-    pub content: String,
-    pub status: String,
-}
 
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct ParsedLogEntry {
-    pub timestamp: String,
-    #[serde(rename = "type")]
-    pub log_type: String,
-    pub content: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub usage: Option<ExecutionUsage>,
-    #[serde(rename = "toolName", alias = "tool_name", skip_serializing_if = "Option::is_none")]
-    pub tool_name: Option<String>,
-    #[serde(rename = "toolInputJson", alias = "tool_input_json", skip_serializing_if = "Option::is_none")]
-    pub tool_input_json: Option<String>,
-}
 
-impl ParsedLogEntry {
-    pub fn new(log_type: impl Into<String>, content: impl Into<String>) -> Self {
-        Self {
-            timestamp: utc_timestamp(),
-            log_type: log_type.into(),
-            content: content.into(),
-            usage: None,
-            tool_name: None,
-            tool_input_json: None,
-        }
-    }
 
-    pub fn info(content: impl Into<String>) -> Self {
-        Self::new("info", content)
-    }
 
-    pub fn error(content: impl Into<String>) -> Self {
-        Self::new("error", content)
-    }
 
-    pub fn stderr(content: impl Into<String>) -> Self {
-        Self::new("stderr", content)
-    }
 
-    pub fn with_usage(mut self, usage: ExecutionUsage) -> Self {
-        self.usage = Some(usage);
-        self
-    }
-}
 
-// Request/Response types
-#[derive(Deserialize, Serialize)]
-pub struct CreateTodoRequest {
-    pub title: String,
-    pub prompt: String,
-    #[serde(default)]
-    pub tag_ids: Vec<i64>,
-    #[serde(default)]
-    pub executor: Option<String>,
-    #[serde(default)]
-    pub scheduler_enabled: Option<bool>,
-    #[serde(default)]
-    pub scheduler_config: Option<String>,
-    #[serde(default)]
-    pub scheduler_timezone: Option<String>,
-    #[serde(default)]
-    pub acceptance_criteria: Option<String>,
-    #[serde(default)]
-    pub webhook_enabled: Option<bool>,
-    #[serde(default)]
-    pub auto_review_enabled: Option<bool>,
-    /// 工作空间 ID（project_directories.id），唯一键。
-    /// 创建时必填；handler 据此查 path 写入 DB cwd 字段。
-    /// 使用 #[serde(default)]：v1 路由从 URL 路径覆盖此值，body 中不传也不影响。
-    #[serde(default)]
-    pub workspace_id: Option<i64>,
-    /// Action 类型标记（如 "rewrite_title"、"optimize_prompt"），
-    /// 仅供前端 ActionButton 组件做 UI 分类展示，不影响执行逻辑。
-    #[serde(default)]
-    pub action_type: Option<String>,
-    /// Action 键值，与 action_type 配合唯一标识一个 action 模板 todo。
-    #[serde(default)]
-    pub action_key: Option<String>,
-    /// 专家/团队名称（WorkBuddy plugin.json 中的 name 字段）。
-    /// 执行时自动加载对应的 Agent MD 和 Skills 注入 prompt。
-    #[serde(default)]
-    pub expert_name: Option<String>,
-    /// 任务级执行模型（覆盖执行器默认）。None = 用执行器默认模型。
-    #[serde(default)]
-    pub model: Option<String>,
-}
 
-#[derive(Deserialize, Serialize)]
-pub struct UpdateTodoRequest {
-    #[serde(default)]
-    pub title: Option<String>,
-    #[serde(default)]
-    pub prompt: Option<String>,
-    #[serde(default)]
-    pub status: Option<TodoStatus>,
-    #[serde(default)]
-    pub executor: Option<String>,
-    #[serde(default)]
-    pub scheduler_enabled: Option<bool>,
-    #[serde(default)]
-    pub scheduler_config: Option<String>,
-    #[serde(default)]
-    pub scheduler_timezone: Option<String>,
-    /// 工作空间 ID（project_directories.id）。
-    /// None=保持当前工作空间，Some(id)=迁移到该工作空间。
-    /// 不接受路径——handler 一律按 id 解析 cwd 路径写入两列。
-    #[serde(default)]
-    pub workspace_id: Option<i64>,
-    #[serde(default)]
-    pub webhook_enabled: Option<bool>,
-    #[serde(default)]
-    pub acceptance_criteria: Option<String>,
-    /// None=不变, Some(true)/Some(false)=更新. 不允许改 reviewer template 的开关.
-    #[serde(default)]
-    pub auto_review_enabled: Option<bool>,
-    /// Action 类型标记（如 "rewrite_title"、"optimize_prompt"），
-    /// 仅供前端 ActionButton 组件做 UI 分类展示，不影响执行逻辑。
-    #[serde(default)]
-    pub action_type: Option<String>,
-    /// Action 键值，与 action_type 配合唯一标识一个 action 模板 todo。
-    #[serde(default)]
-    pub action_key: Option<String>,
-    /// 专家/团队名称（WorkBuddy plugin.json 中的 name 字段）。
-    #[serde(default)]
-    pub expert_name: Option<String>,
-    /// 任务级执行模型（覆盖执行器默认）。None = 不修改；Some("") = 清除；Some(v) = 设置。
-    #[serde(default)]
-    pub model: Option<String>,
-}
 
-#[derive(Deserialize, Serialize)]
-pub struct UpdateTagsRequest {
-    pub tag_ids: Vec<i64>,
-}
 
-/// `PUT /api/todos/{id}/webhook` 请求体：开启/关闭事件驱动。
-///
-/// 扁平具名路由（设计文档），与 `PUT /api/todos/{id}/scheduler` 对称，
-/// 让前端有一个明确的「事件驱动启停」入口，而非塞进通用 update_todo。
-#[derive(Deserialize, Serialize)]
-pub struct UpdateWebhookRequest {
-    pub webhook_enabled: bool,
-}
 
-#[derive(Deserialize, Serialize)]
-pub struct CreateTagRequest {
-    pub name: String,
-    pub color: String,
-}
 
-#[derive(Deserialize, Serialize)]
-pub struct ExecuteRequest {
-    pub todo_id: i64,
-    pub message: Option<String>,
-    pub executor: Option<String>,
-    /// 手动执行时临时指定模型（优先级最高，覆盖 todo.model / executor.default_model）。
-    #[serde(default)]
-    pub model: Option<String>,
-    #[serde(default)]
-    pub params: Option<std::collections::HashMap<String, String>>,
-}
 
-#[derive(Deserialize)]
-pub struct SmartCreateRequest {
-    pub content: String,
-    /// 工作空间 ID（用于查询该工作空间的默认响应 Todo）
-    pub workspace_id: i64,
-}
 
-#[derive(Deserialize)]
-pub struct TodoIdQuery {
-    #[serde(default)]
-    pub todo_id: Option<i64>,
-    #[serde(default)]
-    pub step_id: Option<i64>,
-    #[serde(default)]
-    pub page: Option<i64>,
-    #[serde(default)]
-    pub limit: Option<i64>,
-    #[serde(default)]
-    pub status: Option<String>,
-    /// 按工作空间 ID 过滤；不传则不过滤。当提供 todo_id 或 step_id 时忽略此字段。
-    #[serde(default)]
-    pub workspace_id: Option<i64>,
-    /// 按最近 N 小时过滤（对 finished_at 生效）；不传或 0 表示不过滤。
-    #[serde(default)]
-    pub hours: Option<u32>,
-}
 
-/// 批量更新事项执行器请求体。
-#[derive(Debug, Clone, Deserialize)]
-pub struct BatchUpdateTodoExecutorRequest {
-    pub ids: Vec<i64>,
-    pub executor: String,
-}
 
-/// 批量更新事项执行器返回结果。
-#[derive(Debug, Clone, Serialize)]
-pub struct BatchUpdateTodoResult {
-    pub updated_count: i64,
-    pub total: i64,
-}
 
-/// 批量更新事项工作空间请求体（移动到其他工作空间）。
-#[derive(Debug, Clone, Deserialize)]
-pub struct BatchUpdateTodoWorkspaceRequest {
-    pub ids: Vec<i64>,
-    /// 目标工作空间 ID（project_directories.id）。
-    pub workspace_id: i64,
-}
 
-/// 批量复制事项到其他工作空间请求体。
-#[derive(Debug, Clone, Deserialize)]
-pub struct BatchCopyTodoWorkspaceRequest {
-    pub ids: Vec<i64>,
-    /// 目标工作空间 ID（project_directories.id）。
-    pub workspace_id: i64,
-}
 
-/// 批量暂停/恢复周期执行请求体。scheduler_enabled 为 true 表示恢复，false 表示暂停。
-#[derive(Debug, Clone, Deserialize)]
-pub struct BatchUpdateTodoSchedulerRequest {
-    pub ids: Vec<i64>,
-    pub scheduler_enabled: bool,
-}
 
-/// 批量更新环路工作空间请求体（移动到其他工作空间）。
-#[derive(Debug, Clone, Deserialize)]
-pub struct BatchUpdateLoopWorkspaceRequest {
-    pub ids: Vec<i64>,
-    /// 目标工作空间 ID（project_directories.id）。
-    pub workspace_id: i64,
-}
 
-/// 批量复制环路到其他工作空间请求体。
-#[derive(Debug, Clone, Deserialize)]
-pub struct BatchCopyLoopWorkspaceRequest {
-    pub ids: Vec<i64>,
-    /// 目标工作空间 ID（project_directories.id）。
-    pub workspace_id: i64,
-}
 
-/// 批量 workspace 操作返回结果（通用）。
-#[derive(Debug, Clone, Serialize)]
-pub struct BatchWorkspaceResult {
-    pub updated_count: i64,
-    pub total: i64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RunningBoardResponse {
-    pub records: Vec<ExecutionRecord>,
-    pub scheduled_todos: Vec<Todo>,
-    pub total: i64,
-    pub page: i64,
-    pub limit: i64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecutionRecordsPage {
-    pub records: Vec<ExecutionRecord>,
-    pub total: i64,
-    pub page: i64,
-    pub limit: i64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecutionLogsPage {
-    pub logs: Vec<ParsedLogEntry>,
-    pub total: i64,
-    pub page: i64,
-    pub per_page: i64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecutorCount {
-    pub executor: String,
-    pub count: i64,
-    pub execution_count: i64,
-    pub success_count: i64,
-    pub failed_count: i64,
-    pub total_input_tokens: u64,
-    pub total_output_tokens: u64,
-    pub total_cost_usd: f64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TagCount {
-    pub tag_id: i64,
-    pub tag_name: String,
-    pub tag_color: String,
-    pub count: i64,
-    pub execution_count: i64,
-    pub success_count: i64,
-    pub failed_count: i64,
-    pub total_input_tokens: u64,
-    pub total_output_tokens: u64,
-    pub total_cost_usd: f64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ModelCount {
-    pub model: String,
-    pub count: i64,
-    pub execution_count: i64,
-    pub success_count: i64,
-    pub failed_count: i64,
-    pub total_input_tokens: u64,
-    pub total_output_tokens: u64,
-    pub total_cache_read_tokens: u64,
-    pub total_cache_creation_tokens: u64,
-    pub total_cost_usd: f64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DailyExecution {
-    pub date: String,
-    pub success: i64,
-    pub failed: i64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DailyTokenStats {
-    pub date: String,
-    pub input_tokens: u64,
-    pub output_tokens: u64,
-    pub cache_read_tokens: u64,
-    pub cache_creation_tokens: u64,
-    pub total_cost_usd: f64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DashboardStats {
-    pub total_todos: i64,
-    pub pending_todos: i64,
-    pub running_todos: i64,
-    pub completed_todos: i64,
-    pub failed_todos: i64,
-    pub total_tags: i64,
-    pub scheduled_todos: i64,
-    pub total_executions: i64,
-    pub success_executions: i64,
-    pub failed_executions: i64,
-    pub total_input_tokens: u64,
-    pub total_output_tokens: u64,
-    pub total_cache_read_tokens: u64,
-    pub total_cache_creation_tokens: u64,
-    pub total_cost_usd: f64,
-    pub avg_duration_ms: u64,
-    pub executor_distribution: Vec<ExecutorCount>,
-    pub tag_distribution: Vec<TagCount>,
-    pub model_distribution: Vec<ModelCount>,
-    pub daily_executions: Vec<DailyExecution>,
-    pub daily_token_stats: Vec<DailyTokenStats>,
-    pub recent_executions: Vec<ExecutionRecord>,
-    pub trigger_type_distribution: Vec<TriggerTypeCount>,
-    pub executor_duration_stats: Vec<ExecutorDuration>,
-    pub model_cache_stats: Vec<ModelCacheStat>,
-    // Enhanced metrics
-    pub today_executions: i64,
-    pub executions_change: Option<f64>,
-    pub success_rate_change: Option<f64>,
-    pub cost_change: Option<f64>,
-    pub active_days: i64,
-    pub streak_days: i64,
-    pub peak_daily_executions: i64,
-    pub top_model: Option<String>,
-    pub top_model_tokens: Option<u64>,
-    pub leaderboard: Vec<LeaderboardItem>,
-    // Skills metrics
-    pub skills_stats: Option<SkillsStats>,
-    // Backup metrics
-    pub backup_stats: Option<BackupStats>,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LeaderboardItem {
-    pub rank: i32,
-    pub name: String,
-    pub tokens: u64,
-    pub sessions: i64,
-    pub change: Option<f64>,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RecentCompletedTodo {
-    pub todo_id: i64,
-    pub title: String,
-    pub prompt: Option<String>,
-    pub executor: Option<String>,
-    pub tag_ids: Vec<i64>,
-    /// 所属工作空间（056 补充：纪念板据此反查项目名，免去前端按 id 二次查询）
-    #[serde(default)]
-    pub workspace_id: Option<i64>,
-    pub completed_at: String,
-    pub result: Option<String>,
-    pub model: Option<String>,
-    pub usage: Option<ExecutionUsage>,
-    pub execution_status: String,
-    pub trigger_type: String,
-    pub record_id: i64,
-    /// User-provided score for the most recent execution record (0-100).
-    /// Mirrors `ExecutionRecord::rating` so that the conclusion/memorial view
-    /// can render the score badge without an extra round-trip per card.
-    pub rating: Option<i32>,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TriggerTypeCount {
-    pub trigger_type: String,
-    pub count: i64,
-    pub success_count: i64,
-    pub failed_count: i64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecutorDuration {
-    pub executor: String,
-    pub avg_duration_ms: f64,
-    pub execution_count: i64,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ModelCacheStat {
-    pub model: String,
-    pub total_input_tokens: u64,
-    pub total_cache_read_tokens: u64,
-    pub cache_hit_rate: f64,
-}
 
-// Skills invocation statistics
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SkillsStats {
-    pub total_invocations: i64,
-    pub success_invocations: i64,
-    pub failed_invocations: i64,
-    pub avg_duration_ms: f64,
-    pub invocations_today: i64,
-    pub top_skills: Vec<SkillTop>,
-    pub executor_skills_count: Vec<ExecutorSkillCount>,
-    pub daily_invocations: Vec<DailySkillInvocation>,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SkillTop {
-    pub skill_name: String,
-    pub count: i64,
-    pub success_rate: f64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecutorSkillCount {
-    pub executor: String,
-    pub skills_count: i64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DailySkillInvocation {
-    pub date: String,
-    pub count: i64,
-    pub success: i64,
-}
-
-// Backup statistics
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BackupStats {
-    pub auto_backup_enabled: bool,
-    pub last_backup: Option<String>,
-    pub auto_backup_cron: String,
-    pub database: BackupCategoryStats,
-    pub todo: BackupCategoryStats,
-    pub skills: BackupCategoryStats,
-    pub total_file_count: i64,
-    pub total_size: i64,
-    pub total_size_formatted: String,
-    pub recent_backups: Vec<RecentBackup>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BackupCategoryStats {
-    pub file_count: i64,
-    pub total_size: i64,
-    pub last_backup: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RecentBackup {
-    #[serde(rename = "type")]
-    pub backup_type: String,
-    pub name: String,
-    pub size: i64,
-    pub created_at: String,
-}
 
 #[derive(Deserialize)]
 pub struct UpdateSchedulerRequest {
@@ -1035,123 +160,13 @@ pub mod binding_status {
     pub const RUNNING: &str = "running";
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecutorConfig {
-    pub id: i64,
-    pub name: String,
-    pub path: String,
-    pub enabled: bool,
-    pub display_name: String,
-    pub session_dir: String,
-    /// 是否为系统默认执行器
-    pub is_default: bool,
-    /// 执行器级默认模型。None = 未指定，执行时不传 --model，由执行器配置文件决定。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub default_model: Option<String>,
-    /// 是否支持动态列模型（computed，不落库）。前端据此决定 Select(有选项)/Input(手填)。
-    #[serde(default)]
-    pub supports_models: bool,
-    pub created_at: Option<String>,
-    pub updated_at: Option<String>,
-}
 
-#[derive(Deserialize)]
-pub struct UpdateExecutorRequest {
-    pub path: Option<String>,
-    pub enabled: Option<bool>,
-    pub display_name: Option<String>,
-    pub session_dir: Option<String>,
-    /// 执行器默认模型。空串 = 清除默认模型；None = 不修改。
-    #[serde(default)]
-    pub default_model: Option<String>,
-}
 
-#[derive(Serialize)]
-pub struct ExecutorDetectResult {
-    pub binary_found: bool,
-    pub path_resolved: Option<String>,
-}
 
-/// resolve 操作用的结果：包含检测结果 + 是否触发了数据库更新
-#[derive(Serialize)]
-pub struct ExecutorPathResolveResult {
-    pub binary_found: bool,
-    pub path_resolved: Option<String>,
-    /// 数据库路径是否被更新（仅在 binary_found=true 且路径与原值不同时为 true）
-    pub path_updated: bool,
-    pub old_path: Option<String>,
-    pub new_path: Option<String>,
-}
 
-#[derive(Serialize)]
-pub struct ExecutorTestResult {
-    pub test_passed: bool,
-    pub output: Option<String>,
-    pub error: Option<String>,
-}
 
-#[derive(Serialize)]
-pub struct ExecutorBatchDetectResult {
-    pub results: Vec<ExecutorDetectInfo>,
-    pub total: usize,
-    pub found_count: usize,
-}
 
-#[derive(Serialize)]
-pub struct ExecutorDetectInfo {
-    pub name: String,
-    pub display_name: String,
-    pub binary_found: bool,
-    pub path_resolved: Option<String>,
-    pub enabled: bool,
-}
 
-// Executor types
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Hash)]
-#[serde(rename_all = "lowercase")]
-#[derive(Default)]
-pub enum ExecutorType {
-    Mobilecoder,
-    #[default]
-    Claudecode,
-    Codebuddy,
-    Opencode,
-    Atomcode,
-    Hermes,
-    Kimi,
-    Codex,
-    Codewhale,
-    Pi,
-    Mimo,
-    Zhanlu,
-    Kilo,
-}
-
-impl ExecutorType {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            ExecutorType::Mobilecoder => "mobilecoder",
-            ExecutorType::Claudecode => "claudecode",
-            ExecutorType::Codebuddy => "codebuddy",
-            ExecutorType::Opencode => "opencode",
-            ExecutorType::Atomcode => "atomcode",
-            ExecutorType::Hermes => "hermes",
-            ExecutorType::Kimi => "kimi",
-            ExecutorType::Codex => "codex",
-            ExecutorType::Codewhale => "codewhale",
-            ExecutorType::Pi => "pi",
-            ExecutorType::Mimo => "mimo",
-            ExecutorType::Zhanlu => "zhanlu",
-            ExecutorType::Kilo => "kilo",
-        }
-    }
-}
-
-impl std::fmt::Display for ExecutorType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.as_str())
-    }
-}
 
 // Unified API Response
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -1227,75 +242,11 @@ pub struct UpdateReviewTemplateRequest {
     pub prompt: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TodoTemplate {
-    pub id: i64,
-    pub title: String,
-    pub prompt: Option<String>,
-    pub category: String,
-    pub sort_order: i32,
-    pub is_system: bool,
-    pub source_url: Option<String>,
-    pub last_sync_at: Option<String>,
-    pub created_at: Option<String>,
-    pub updated_at: Option<String>,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CreateTemplateRequest {
-    pub title: String,
-    pub prompt: Option<String>,
-    pub category: String,
-    pub sort_order: Option<i32>,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UpdateTemplateRequest {
-    #[serde(default)]
-    pub title: Option<String>,
-    #[serde(default)]
-    pub prompt: Option<String>,
-    #[serde(default)]
-    pub category: Option<String>,
-    #[serde(default)]
-    pub sort_order: Option<i32>,
-}
 
-/// 导入导出备份数据
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BackupData {
-    pub version: String,
-    pub created_at: String,
-    pub tags: Vec<TagBackup>,
-    pub todos: Vec<TodoBackup>,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TagBackup {
-    pub name: String,
-    pub color: String,
-}
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TodoBackup {
-    pub title: String,
-    pub prompt: String,
-    pub status: TodoStatus,
-    pub executor: Option<String>,
-    pub scheduler_enabled: bool,
-    pub scheduler_config: Option<String>,
-    pub tag_names: Vec<String>,
-    pub workspace_path: Option<String>,
-    pub worktree: Option<String>,
-    pub action_type: Option<String>,
-    pub action_key: Option<String>,
-    /// 备份时的工作空间 ID，为空表示未分配
-    #[serde(default)]
-    pub workspace_id: Option<i64>,
-    /// 任务级指定的执行模型（备份恢复后保留，不影响老备份导入）。
-    #[serde(default)]
-    pub model: Option<String>,
-}
 
 // ============ 环路导入导出 DTO ============
 // 方案：伪ID引用（@loop_1, @todo_1 等）解决跨实体引用问题
@@ -1314,70 +265,12 @@ pub mod codes {
 
 // ============ 伪ID工具函数 ============
 
-/// 伪ID类型前缀
-const PSEUDO_ID_PREFIXES: &[&str] = &["loop", "todo", "step", "trigger", "template", "tag"];
 
-/// 生成伪ID: "@{prefix}_{index}"
-pub fn generate_pseudo_id(prefix: &str, index: usize) -> String {
-    format!("@{}_{}", prefix, index)
-}
 
-/// 校验伪ID格式是否合法
-/// 格式: ^@(loop|todo|step|trigger|template|tag)_\d+$
-pub fn validate_pseudo_id(id: &str) -> bool {
-    // 必须以 @ 开头
-    if !id.starts_with('@') {
-        return false;
-    }
-    // 去掉 @ 前缀后检查格式
-    let rest = &id[1..];
-    // 检查是否包含 _ 和数字部分
-    if let Some(underscore_pos) = rest.find('_') {
-        let prefix = &rest[..underscore_pos];
-        let suffix = &rest[underscore_pos + 1..];
-        return PSEUDO_ID_PREFIXES.contains(&prefix) && suffix.parse::<usize>().is_ok();
-    }
-    false
-}
 
-/// 从伪ID提取前缀
-pub fn extract_pseudo_prefix(id: &str) -> Option<&str> {
-    if !id.starts_with('@') {
-        return None;
-    }
-    let rest = &id[1..];
-    // 没有下划线则不是合法伪ID格式
-    let underscore_pos = rest.find('_')?;
-    Some(&rest[..underscore_pos])
-}
 
-/// 从伪ID提取序号
-pub fn extract_pseudo_index(id: &str) -> Option<usize> {
-    if !id.starts_with('@') {
-        return None;
-    }
-    let rest = &id[1..];
-    rest.split('_').nth(1)?.parse().ok()
-}
 
-/// 返回当前 UTC 时间的 ISO 8601 格式字符串 (2024-01-15T08:30:00.000Z)
-pub fn utc_timestamp() -> String {
-    chrono::Utc::now().format("%Y-%m-%dT%H:%M:%S%.3fZ").to_string()
-}
 
-/// 093：返回「hours 小时前」的 UTC cutoff，格式与 `utc_timestamp` 完全一致。
-///
-/// 用途：列表 hours 过滤从 `REPLACE(REPLACE(updated_at,...)>= datetime('now',...)`
-/// （列上套函数使索引失效）改为 `updated_at >= ?` 参数绑定裸列比较。
-/// 存储侧格式考据（生产统一 T/Z ISO，见 093 设计文档 §1.3）：
-/// 应用层写入走 `utc_timestamp()`（毫秒精度），触发器兜底走 `%Y-%m-%dT%H:%M:%SZ`
-/// （秒级精度）；两种形态与本 cutoff 做字符串比较在秒级等价于时间比较，
-/// 毫秒边界误差 <1s，对 hours 级过滤可忽略。
-pub fn utc_timestamp_minus_hours(hours: u32) -> String {
-    (chrono::Utc::now() - chrono::Duration::hours(i64::from(hours)))
-        .format("%Y-%m-%dT%H:%M:%S%.3fZ")
-        .to_string()
-}
 
 #[cfg(test)]
 #[allow(clippy::unwrap_used, clippy::expect_used, clippy::panic, clippy::useless_vec, clippy::redundant_pattern_matching, clippy::redundant_clone, clippy::len_zero, clippy::bool_assert_comparison, clippy::unnecessary_get_then_check, clippy::doc_lazy_continuation, clippy::clone_on_copy, clippy::print_stdout, clippy::needless_pass_by_value, clippy::sliced_string_as_bytes, clippy::manual_map, clippy::collapsible_match, clippy::question_mark)]
@@ -1788,63 +681,7 @@ mod tests {
     }
 }
 
-/// Replace placeholders in a string using a map of key-value pairs.
-/// Format: `{{key}}` will be replaced with the corresponding value from the map.
-/// If a key is not found in the map, it remains unchanged.
-///
-/// **Footgun — value 中含占位符**: 如果某个 `value` 本身包含 `{{otherkey}}` 而
-/// `otherkey` 也在 `params` 里,**只有当 `otherkey` 先于当前 (k,v) 被替换时**,value
-/// 中的 `{{otherkey}}` 才会被吃掉。`HashMap` 的迭代顺序是 `RandomState` 加盐的随机化,
-/// 因此这种行为不可预测。**调用方请避免在 value 中嵌入另一个 key 的占位符**,或
-/// 自行预处理 value(把嵌入的占位符先替换为最终文本)。
-pub fn replace_placeholders(text: &str, params: &std::collections::HashMap<String, String>) -> String {
-    let mut result = text.to_string();
-    for (key, value) in params {
-        let placeholder = format!("{{{{{}}}}}", key);
-        result = result.replace(&placeholder, value);
-    }
-    result
-}
 
-/// Build standard trigger params from message content.
-/// This unifies how params are constructed across slash commands, default responses,
-/// and other trigger types.
-///
-/// Returns (trigger_type, params):
-/// - For slash commands (content starts with '/'): trigger_type = "slash_command"
-/// - For other messages: trigger_type = "default_response"
-///
-/// Standard params always include:
-/// - `content`: the message body
-/// - `message`: the message body
-/// - `raw_message`: full raw message (for slash commands, includes the command prefix)
-pub fn build_trigger_params(content: &str) -> (String, std::collections::HashMap<String, String>) {
-    let trimmed = content.trim();
-
-    if trimmed.starts_with('/') {
-        let mut parts = trimmed.splitn(2, char::is_whitespace);
-        let command = parts.next().unwrap_or("").trim();
-        let body = parts.next().unwrap_or("").trim();
-
-        if !body.is_empty() {
-            let mut params = std::collections::HashMap::new();
-            params.insert("content".to_string(), body.to_string());
-            params.insert("message".to_string(), body.to_string());
-            params.insert(
-                "raw_message".to_string(),
-                format!("{} {}", command, body).trim().to_string(),
-            );
-            params.insert("slash_command".to_string(), command.to_string());
-            return ("slash_command".to_string(), params);
-        }
-    }
-
-    let mut params = std::collections::HashMap::new();
-    params.insert("content".to_string(), trimmed.to_string());
-    params.insert("message".to_string(), trimmed.to_string());
-    params.insert("raw_message".to_string(), trimmed.to_string());
-    ("default_response".to_string(), params)
-}
 
 #[cfg(test)]
 #[allow(clippy::unwrap_used, clippy::expect_used, clippy::panic, clippy::useless_vec, clippy::redundant_pattern_matching, clippy::redundant_clone, clippy::len_zero, clippy::bool_assert_comparison, clippy::unnecessary_get_then_check, clippy::doc_lazy_continuation, clippy::clone_on_copy, clippy::print_stdout, clippy::needless_pass_by_value, clippy::sliced_string_as_bytes, clippy::manual_map, clippy::collapsible_match, clippy::question_mark)]
